Fix signed/unsigned warnings, tolower UB, and missing npos check
- conv.hh: change ~15 loop variables from auto/int to size_t
- variable_name_utils.hh: cast char to unsigned char before tolower()
  to avoid undefined behavior with negative char values
- util.hh: add npos check in remove_file_extension(), handle snprintf
  error in series_filename(), fix int vs size_t in glob()
- lattice.hh: fix remaining int vs size_t in reshape(vector<int>)

+++ include/ndarray/lattice.hh
@@ -162,7 +162,7 @@ inline void lattice::reshape(const std::vector<int> &sizes)
 {
   // Convert int to size_t
   std::vector<size_t> _sizes(sizes.size());
-  for(int i = 0; i < sizes.size(); ++i) {
+  for(size_t i = 0; i < sizes.size(); ++i) {
     _sizes[i] = sizes[i];
   }
